src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/drag-and-drop-in-shadow-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests whether we can start dragging a node in shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
